+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### Changed
 
-- NuvlaBox Cluster - improved cluster update workflow, with added status notes for clarity
+- NuvlaBox Cluster - improved cluster update workflow, with added status
+  notes for clarity.
+- Kafka - Allow loading Kafka producer conditionally based on
+  KAFKA_PRODUCER_INIT env var. Added an option to configure Kafka client from
+  env vars. Bumped up Kafka client's version to 0.1.25.
+- Elasticsearch - load sniffer conditionally based on env var ES_SNIFFER_INIT.
 
 ## [5.23.2] - 2022-02-04