+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+- NuvlaBox Cluster - improved cluster update workflow, with added status
+  notes for clarity.
 - Kafka - Allow loading Kafka producer conditionally based on
   KAFKA_PRODUCER_INIT env var. Added an option to configure Kafka client from
   env vars. Bumped up Kafka client's version to 0.1.25.