+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 
 ### Changed
 
+- Cleaned up dead code server
+  start [#648](https://github.com/nuvla/api-server/issues/648)
 - Two factor authentication - Bigger email token (6 digits)
-- NuvlaBox Cluster - improved cluster update workflow, with added status
-  notes for clarity.
+- NuvlaBox Cluster - improved cluster update workflow, with added status notes
+  for clarity.
 - Kafka - Allow loading Kafka producer conditionally based on
   KAFKA_PRODUCER_INIT env var. Added an option to configure Kafka client from
   env vars. Bumped up Kafka client's version to 0.1.25.